requirements.dev.txt
boto3

# for unit tests
pytest
pyyaml
